docs: Renamed AuxillarySensors and NetworkDispatch - Renamed AuxillarySensors.tsx to AuxillarySensorStore.tsx for   consistency - Renamed NetworkDispatch.tsx to NetworkDispatchStore.tsx for   consistency - Updated necessary imports with new names
